+++ sema/src/infer/functions.rs
@@ -15,6 +15,11 @@ impl TypeInference {
             _ => (None, None),
         };
 
+        let saved_type_params = std::mem::replace(
+            &mut self.type_params_in_scope,
+            func.type_params.clone(),
+        );
+
         for type_param in &func.type_params {
             let fresh_var = self.type_gen.fresh();
             self.env.define_local(type_param.clone(), fresh_var);
@@ -80,6 +85,7 @@ impl TypeInference {
 
         self.pop_return_type();
         self.env = saved_env;
+        self.type_params_in_scope = saved_type_params;
 
         TypedFunction {
             name: func.name.clone(),